django-api/requirements.txt
django==4.2.21
djangorestframework==3.15.2
drf-yasg==1.21.10
djangorestframework-simplejwt==5.3.1
django-filter==24.3
pillow==10.4.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
gunicorn
django-storages==1.14.6
django-environ==0.12.0
boto3==1.38.26